Fix all warnings.
git-svn-id: http://svn.inebriatedinnovations.com/repo/BrewBit/sw/b2tc-ChibiOS@585 ad39a8c3-5e57-4819-ba6d-ac3c945a73f6

+++ src/gui_home.c
@@ -128,8 +128,8 @@ home_screen_msg(msg_event_t* event)
   home_screen_t* s = widget_get_instance_data(event->widget);
   if (event->msg_id == MSG_NEW_TEMP) {
     temp_msg_t* msg = event->msg_data;
-    sprintf(s->temp_str1, "%d", msg->temp);
-    sprintf(s->temp_str2, "%d", msg->temp);
+    sprintf(s->temp_str1, "%d", (int)msg->temp);
+    sprintf(s->temp_str2, "%d", (int)msg->temp);
     widget_invalidate(s->probe1_temp_label);
     widget_invalidate(s->probe2_temp_label);
   }
